Error handling and checks before storing data into local/session storage
Service was down for API so I added some error handling.
It was previously adding an empty stringify array into local storage and ruining the functionality, so if the fetch returns an array of length 0, it will no longer add it to the storage.

Also, if the service is down or something goes wrong with fetching the API, a window alert will appear, letting the user know and suggest trying later.